feat: move all to parsers folder
