+++ src/state/apis/zapper/validators.ts
@@ -1,4 +1,4 @@
-import type { AssetId, AssetNamespace, ChainId } from '@shapeshiftoss/caip'
+import type { AssetId, ChainId } from '@shapeshiftoss/caip'
 import {
   ASSET_NAMESPACE,
   avalancheChainId,
@@ -12,8 +12,6 @@ import { invert } from 'lodash'
 import type { Infer, Type } from 'myzod'
 import z from 'myzod'
 import { isNonEmpty, isUrl } from 'lib/utils'
-
-import type { NftCollectionType, NftItem } from '../nft/types'
 
 export enum SupportedZapperNetwork {
   Avalanche = 'avalanche',
@@ -854,43 +852,6 @@ const V2AppResponse = z.object({
   supportedNetworks: z.array(V2AppSupportedNetworkResponse),
   groups: z.array(V2AppGroupResponse),
 })
-
-export const parseToNftItem = (zapperItem: V2NftUserItem, chainId: ChainId): NftItem => {
-  const {
-    token,
-    token: { lastSaleEth, medias, tokenId },
-    token: { collection },
-  } = zapperItem
-
-  const collectionItem: NftCollectionType = {
-    id: collection?.address
-      ? toAssetId({
-          assetReference: collection.address,
-          assetNamespace: collection.nftStandard as AssetNamespace,
-          chainId,
-        })
-      : '',
-    chainId,
-    description: '', // Not supported by the /v2/nft/user/tokens endpoint
-    name: collection?.name || '',
-    floorPrice: collection?.floorPriceEth || '',
-    openseaId: collection?.openseaId || '',
-    socialLinks: [], // Not supported by the /v2/nft/user/tokens endpoint
-  }
-
-  const nftItem: NftItem = {
-    id: tokenId,
-    name: token.name,
-    description: null, // Not supported by the /v2/nft/user/tokens endpoint
-    price: lastSaleEth || '',
-    chainId,
-    collection: collectionItem,
-    medias: medias || [],
-    rarityRank: token.rarityRank,
-  }
-
-  return nftItem
-}
 
 export const V2AppsResponse = z.array(V2AppResponse)
 export type V2AppResponseType = Infer<typeof V2AppResponse>
